+++ apps/zui/src/css/_utilities.scss
@@ -191,3 +191,7 @@
 .overflow-hidden {
   overflow: hidden;
 }
+
+.hidden {
+  display: none;
+}
