+++ apps/zui/src/css/settings/_variables.scss
@@ -37,4 +37,14 @@
 
 :root {
   --gutter-space: var(--space-s);
+  --hover-duration: 75ms;
+
+  /* Hover and Active Filters */
+  --hover-filter: brightness(95%);
+  --active-filter: brightness(93%);
+
+  @media (prefers-color-scheme: dark) {
+    --hover-filter: brightness(125%);
+    --active-filter: brightness(135%);
+  }
 }
