+++ apps/zui/src/css/utilities/_flex.scss
@@ -1,6 +1,11 @@
 .justify-between {
   justify-content: space-between;
 }
+
 .items-start {
   align-items: flex-start;
+}
+
+.items-center {
+  align-items: center;
 }
