+++ apps/zui/src/views/settings-modal/section.tsx
@@ -1,9 +1,9 @@
-import {isArray} from "lodash"
-import {useSelector} from "react-redux"
-import {useDispatch} from "src/app/core/state"
 import {ConfigItem} from "src/domain/configurations/plugin-api"
-import ConfigPropValues from "src/js/state/ConfigPropValues"
 import {Config} from "src/zui"
+import {Input} from "./input"
+import Link from "src/js/components/common/Link"
+
+export type SettingProps = {field: ConfigItem; sectionName: string}
 
 export function Section(props: {config: Config}) {
   return (
@@ -21,86 +21,20 @@ export function Section(props: {config: Config}) {
   )
 }
 
-type SettingProps = {field: ConfigItem; sectionName: string}
-
 function Setting(props: SettingProps) {
   return (
     <>
-      <label>{props.field.label}</label>
+      <label htmlFor={props.field.name} className="repel">
+        {props.field.label}
+        <HelpLink {...props} />
+      </label>
       <Input {...props} />
     </>
   )
 }
 
-function Input(props: SettingProps) {
-  const dispatch = useDispatch()
-  const {field} = props
-  const {name, defaultValue} = field
-  const value = useSelector(ConfigPropValues.get(props.sectionName, field.name))
-  const update = (value) =>
-    dispatch(
-      ConfigPropValues.set({
-        configName: props.sectionName,
-        propName: field.name,
-        value,
-      })
-    )
-  const onChange = (e) => update(e.currentTarget.value)
-
-  switch (field.type) {
-    case "char":
-      return (
-        <input
-          type="input"
-          style={{inlineSize: "8ch"}}
-          id={name}
-          name={name}
-          onChange={onChange}
-          defaultValue={value === undefined ? defaultValue : value}
-        />
-      )
-    case "file":
-      return (
-        <input
-          type="file"
-          id={name}
-          name={name}
-          onChange={onChange}
-          defaultValue={value === undefined ? defaultValue : value}
-          placeholder="default"
-        />
-      )
-    case "string":
-      if (field.enum) {
-        return (
-          <select
-            id={name}
-            name={name}
-            defaultValue={value === undefined ? defaultValue : value}
-            onChange={onChange}
-          >
-            {field.enum.map((e) => {
-              const label = isArray(e) ? e[0] : e
-              const value = isArray(e) ? e[1] : e
-              return (
-                <option key={value} value={value}>
-                  {label}
-                </option>
-              )
-            })}
-          </select>
-        )
-      } else {
-        return (
-          <input
-            id={name}
-            name={name}
-            onChange={onChange}
-            type="text"
-            placeholder=""
-            defaultValue={value === undefined ? defaultValue : value}
-          />
-        )
-      }
-  }
+function HelpLink(props: SettingProps) {
+  if (!props.field.helpLink) return null
+  const {url, label} = props.field.helpLink
+  return <Link href={url}>{label}</Link>
 }
